pyproject.toml
[tool.poetry]
name = "pokemon"
version = "0.1.0"
description = ""
authors = ["yunti <yunti9@icloud.com>"]

[tool.poetry.dependencies]
python = "^3.8"
djangorestframework = "^3.12.1"
psycopg2-binary = "^2.8.6"
django-environ = "^0.4.5"
Django = "^3.1.3"
django-cors-headers = "^3.5.0"
djangorestframework-simplejwt = "^4.5.0"
requests = "^2.25.0"

[tool.poetry.dev-dependencies]
black = "^20.8b1"
httpie = "^2.3.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
